template/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pyperclip>=1.8.2
pydantic>=2.0.0
anthropic>=0.49.0
jsonschema>=4.23.0
jinja2>=3.1.0
pyyaml>=6.0
